+++ Tests/Unit/Domain/Repository/Registration/RegistrationRepositoryTest.php
@@ -1,0 +1,46 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OliverKlee\Seminars\Tests\Unit\Domain\Repository\Registration;
+
+use Nimut\TestingFramework\TestCase\UnitTestCase;
+use OliverKlee\Oelib\Domain\Repository\Interfaces\DirectPersist;
+use OliverKlee\Seminars\Domain\Repository\Registration\RegistrationRepository;
+use TYPO3\CMS\Extbase\Object\ObjectManagerInterface;
+use TYPO3\CMS\Extbase\Persistence\Repository;
+
+/**
+ * @covers \OliverKlee\Seminars\Domain\Repository\Registration\RegistrationRepository
+ */
+final class RegistrationRepositoryTest extends UnitTestCase
+{
+    /**
+     * @var RegistrationRepository
+     */
+    private $subject;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $objectManagerStub = $this->createMock(ObjectManagerInterface::class);
+        $this->subject = new RegistrationRepository($objectManagerStub);
+    }
+
+    /**
+     * @test
+     */
+    public function isRepository(): void
+    {
+        self::assertInstanceOf(Repository::class, $this->subject);
+    }
+
+    /**
+     * @test
+     */
+    public function implementsDirectPersist(): void
+    {
+        self::assertInstanceOf(DirectPersist::class, $this->subject);
+    }
+}